Create new component for custom dropdown
