+++ usecase-compass-ui/src/components/Form.jsx
@@ -1,8 +1,9 @@
-import { useState, useEffect, useRef } from "react";
+import { useState } from "react";
 import PreviewDialog from "../utils/PreviewDialog";
 import http from "../services/http";
 import AlertDialog from "../utils/AlertDialog";
 import Datepicker from "../utils/DatePicker"
+import Dropdown from "../utils/Dropdown";
 
 const formConfig = [
     { label: 'First Name', name: 'firstNamePrimary', type: 'text', placeholder: 'Enter First Name', errorMessage: 'First name is required', max: 100 },
@@ -41,26 +42,7 @@ const Form = ({ showSuccessMsg }) => {
     const [errors, setErrors] = useState({});
     const [invalidFormError, setInvalidFormError] = useState("");
     const [showFormClearMsg, setShowFormClearMsg] = useState(false);
-    const today = new Date().toISOString().split("T")[0];
     const clearFormConfMsg = { title: "Clear Form", message: "Are you sure you want to clear the form?" };
-    const [showGenderPopup, setShowGenderPopup] = useState(false);
-    const dropdownRef = useRef(null);
-    const dropdownBtnRef = useRef(null);
-
-    useEffect(() => {
-        function handleClickOutside(event) {
-            if (dropdownRef.current && !dropdownRef.current.contains(event.target) && !dropdownBtnRef.current.contains(event.target)) {
-                setShowGenderPopup(false);
-            }
-        }
-    
-        document.addEventListener("mousedown", handleClickOutside);
-    
-        return () => {
-            document.removeEventListener("mousedown", handleClickOutside);
-        };
-    }, [dropdownRef]);
-
 
     const handleChange = (e) => {
         const { name, value, files } = e.target;
@@ -173,7 +155,6 @@ const Form = ({ showSuccessMsg }) => {
             ...prev,
             gender: val,
         }));
-        setShowGenderPopup(false)
     };
 
     const handlePreview = () => {
@@ -206,27 +187,7 @@ const Form = ({ showSuccessMsg }) => {
 
                         {field.type === 'select' ? (
                             <>
-                                <div className="relative inline-block text-left w-full text-[18px]">
-                                    <div>
-                                        <button ref={dropdownBtnRef} onClick={() => setShowGenderPopup((prevState) => !prevState)} type="button" className={`inline-flex w-full h-[60px] items-center justify-between rounded-md border-2 border-[#707070] bg-white px-4 py-2 font-medium shadow-smfocus:outline-none cursor-pointer ${formData.gender ? "text-[#1B2142]" : "text-[#9FA1AD]"}`}>
-                                            {formData.gender ?  formData.gender : field.placeholder}
-                                            <svg className="-mr-1 ml-2 h-5 w-5" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor" aria-hidden="true">
-                                                <path fillRule="evenodd" d="M5.23 7.21a.75.75 0 011.06.02L10 11.293l3.71-4.063a.75.75 0 111.08 1.04l-4.25 4.65a.75.75 0 01-1.08 0l-4.25-4.65a.75.75 0 01.02-1.06z" clipRule="evenodd" />
-                                            </svg>
-                                        </button>
-                                    </div>
-                                    {showGenderPopup && <div ref={dropdownRef} className="absolute z-10 mt-2 px-2 py-1 w-full origin-top-right rounded-md bg-white shadow-lg border-1 border-[#E4E7EC]">
-                                        <div className="py-1" role="menu" aria-orientation="vertical" aria-labelledby="options-menu">
-                                            {field.options.map((eachOpt) => (
-                                                <p onClick={() => setGenderValue(eachOpt)} className={`text-[#101828] px-3 rounded-md py-1 mb-2 flex justify-between items-center ${formData.gender === eachOpt ? "bg-[#F9FAFB]" : ""} cursor-pointer`} key={eachOpt}>
-                                                    <span>{eachOpt}</span>
-                                                   {(formData.gender && formData.gender === eachOpt) && (<img src="/assets/icons/check.svg" alt="check"/>)}
-                                                </p>
-                                            ))}
-                                        </div>
-                                    </div>}
-                                </div>
-
+                                <Dropdown field={field} setValue={setGenderValue}/>
                                 {errors[field.name] && (
                                     <p className="text-red-500 text-xs mt-1">{errors[field.name]}</p>
                                 )}
